+++ Eat_and_Greet/src/comps/ChooseImageButton/ChooseImageButton.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 interface ChooseImageProps {
   onPress?: () => void;
   buttonSize: number;
-  imagePath: string;
+  imagePath?: string;
   color?: string;
 }
